06-14-2018
many MANY bugfixes

added in jump as a method in UnifiedController

Overhauled player control so a single class (PlayerControl) acts as a performable and a decisionMaker
Playercontrol automatically sets up input and splitscreen for the connected players

the _PlayerControl component requires a mind to override decisions for. It continuously checks for an
availible controller until it finds one, then it stays that controller unless manually changed.

Added in the concept of a combo move, added in health and taking away health, added in 3 levels of knockback,
and the concept of a combo ending with big knockback

added a short dash to melee attacks to ensure they connected

increased collider radius of hand and foot to ensure attacks connect