todo: 	-make rigidbody control rotation instead of navagent 	-transition abilities away from ienumerators 	-make PrefabPool return an instance rather than the actual prefab 	-make PrefabPool hold copies of instantiated prefabs so theres no need to go creating and destroying them all the time 	-generalize range attacks to thrownprojectiles and beam projectiles 	-make jumping and climbing not just teleport directly to the apropriate spot 	-make kunai apply effects when colliding with a hitbox as whel as a physical object
------------------------------------------------------------------------------------------------

fixed bug where avatar definition affected jumping ability
changed jump velocity calculation so it now scales along with transform
changed unified controller so rigidbody is synched in the FixedUpdate physics loop rather than the framerate depandant Update loop.
Added in club and added many materials and visual upgrades. This was actually done in the previous version but was undocumented. Final styling of the main room was done in this update