todo: 	-make rigidbody control rotation instead of navagent 	-transition abilities away from ienumerators 	-make PrefabPool return an instance rather than the actual prefab 	-make PrefabPool hold copies of instantiated prefabs so theres no need to go creating and destroying them all the time 	-generalize range attacks to thrownprojectiles and beam projectiles 	-make jumping and climbing not just teleport directly to the apropriate spot 	-make kunai apply effects when colliding with a hitbox as whel as a physical object
------------------------------------------------------------------------------------------------

completed trackedObject and NavMEshSurfaceOther. Both are now fully operational, but mostly untested. So far utilizing these precludes the possibility of any other navMeshAreas because they completely overwrite the nav meesh area mask. This isnt neccessairily a requirement of the components and is more my not understanding bit shift operations and & | ~ operators with ints. THeoretically compatability with other navmeshareas can be added, with caveats. nested areas are also theoretically possible