todo: 	-make rigidbody control rotation instead of navagent 	-transition abilities away from ienumerators 	-make PrefabPool return an instance rather than the actual prefab 	-make PrefabPool hold copies of instantiated prefabs so theres no need to go creating and destroying them all the time 	-generalize range attacks to thrownprojectiles and beam projectiles 	-make jumping and climbing not just teleport directly to the apropriate spot 	-make kunai apply effects when colliding with a hitbox as well as a physical object
------------------------------------------------------------------------------------------------
fixed some naming rule violations
simplified some object initializations

added in the modularnavmesh class
this class bakes a navmesh based on the triggers used as areas. Currently, the "nonwalkable" nav mesh is the built in default navmesharea "jump". this is because of many errors in the verson. These errors are recorded in the "error.txt" document in this folder. The tecnique is proven functional. The next step is baking the navmeshdata directly into the prefab
add automatically adding navmeshlinks